# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CollectionRunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectionRunView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CollectionRunPage",
        "type": "object"
      },
      "CollectionRunRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "ingestion_mode": {
            "enum": [
              "live",
              "backfill"
            ],
            "title": "Ingestion Mode",
            "type": "string"
          },
          "operation": {
            "const": "search_posts",
            "default": "search_posts",
            "title": "Operation",
            "type": "string"
          },
          "policy_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Policy Version",
            "type": "string"
          },
          "request_value": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Value",
            "type": "string"
          },
          "retention_days": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "source": {
            "enum": [
              "x",
              "bilibili",
              "weibo",
              "xiaohongshu",
              "douyin",
              "bluesky"
            ],
            "title": "Source",
            "type": "string"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "source",
          "request_value",
          "since",
          "until",
          "policy_version",
          "retention_days",
          "ingestion_mode"
        ],
        "title": "CollectionRunRequest",
        "type": "object"
      },
      "CollectionRunView": {
        "properties": {
          "budget_day": {
            "format": "date",
            "title": "Budget Day",
            "type": "string"
          },
          "bytes_count": {
            "title": "Bytes Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fencing_token": {
            "title": "Fencing Token",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingestion_mode": {
            "enum": [
              "live",
              "backfill"
            ],
            "title": "Ingestion Mode",
            "type": "string"
          },
          "items_count": {
            "title": "Items Count",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "monitor_version_id": {
            "format": "uuid",
            "title": "Monitor Version Id",
            "type": "string"
          },
          "operation": {
            "enum": [
              "search_posts",
              "fetch_post",
              "list_comments",
              "list_replies"
            ],
            "title": "Operation",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "enum": [
                  "ok",
                  "empty",
                  "partial",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "pages_count": {
            "title": "Pages Count",
            "type": "integer"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "request_value": {
            "title": "Request Value",
            "type": "string"
          },
          "reserved_requests": {
            "title": "Reserved Requests",
            "type": "integer"
          },
          "retention_days": {
            "title": "Retention Days",
            "type": "integer"
          },
          "schedule_slot": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Slot"
          },
          "source": {
            "enum": [
              "x",
              "bilibili",
              "weibo",
              "xiaohongshu",
              "douyin",
              "bluesky"
            ],
            "title": "Source",
            "type": "string"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "trigger": {
            "enum": [
              "manual",
              "scheduled"
            ],
            "title": "Trigger",
            "type": "string"
          },
          "window_since": {
            "format": "date-time",
            "title": "Window Since",
            "type": "string"
          },
          "window_until": {
            "format": "date-time",
            "title": "Window Until",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_id",
          "parent_run_id",
          "monitor_version_id",
          "source",
          "operation",
          "request_value",
          "retention_days",
          "trigger",
          "ingestion_mode",
          "schedule_slot",
          "budget_day",
          "reserved_requests",
          "state",
          "outcome",
          "fencing_token",
          "pages_count",
          "items_count",
          "bytes_count",
          "stop_reason",
          "window_since",
          "window_until",
          "created_at",
          "completed_at"
        ],
        "title": "CollectionRunView",
        "type": "object"
      },
...
      "ContentWithdrawalInput": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "enum": [
              "deleted",
              "purpose_revoked"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ContentWithdrawalInput",
        "type": "object"
      },
      "ContentWithdrawalView": {
        "properties": {
          "affected_knowledge_entries": {
            "title": "Affected Knowledge Entries",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "visibility": {
            "enum": [
              "unavailable",
              "deleted"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "visibility",
          "affected_knowledge_entries"
        ],
        "title": "ContentWithdrawalView",
        "type": "object"
      },
...
      "ErrorView": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id"
        ],
        "title": "ErrorView",
        "type": "object"
      },
...
      "InboxItem": {
        "properties": {
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "post",
              "comment",
              "reply"
            ],
            "title": "Kind",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "monitor_titles": {
            "items": {
              "type": "string"
            },
            "title": "Monitor Titles",
            "type": "array"
          },
          "parent_external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent External Id"
          },
          "provider_namespace": {
            "title": "Provider Namespace",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "relation_status": {
            "enum": [
              "root",
              "unresolved",
              "resolved"
            ],
            "title": "Relation Status",
            "type": "string"
          },
          "reply_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply Count"
          },
          "root_external_id": {
            "title": "Root External Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "x",
              "bilibili",
              "weibo",
              "xiaohongshu",
              "douyin",
              "bluesky"
            ],
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "source",
          "provider_namespace",
          "external_id",
          "kind",
          "root_external_id",
          "parent_external_id",
          "relation_status",
          "text",
          "version",
          "canonical_url",
          "published_at",
          "first_seen_at",
          "last_seen_at",
          "reply_count",
          "monitor_titles"
        ],
        "title": "InboxItem",
        "type": "object"
      },
      "InboxPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InboxItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "InboxPage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OwnerSession": {
        "in": "cookie",
        "name": "hk_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-owner learning workspace. Unsafe requests require an exact allowed Origin; authenticated writes also require a session-bound X-CSRF-Token. Monitor configuration history and zero-network query previews are available; collection remains disabled until a source passes rights and pipeline admission.",
    "title": "HotKey API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/collection-runs": {
      "get": {
        "operationId": "listCollectionRuns",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionRunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "List Collection Runs",
        "tags": [
          "collection"
        ]
      }
    },
    "/api/collection-runs/{identity}": {
      "get": {
        "operationId": "getCollectionRun",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Get Collection Run",
        "tags": [
          "collection"
        ]
      }
    },
    "/api/contents": {
      "get": {
        "operationId": "listInboxContents",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Inbox Contents",
        "tags": [
          "contents"
        ]
      }
    },
    "/api/contents/{identity}/withdraw": {
      "post": {
        "operationId": "withdrawContent",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentWithdrawalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentWithdrawalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Withdraw Content",
        "tags": [
          "contents"
        ]
      }
    },
...
    "/api/monitors/{identity}/runs": {
      "post": {
        "operationId": "createCollectionRun",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[a-zA-Z0-9_.:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "State conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Rate limited"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Create Collection Run",
        "tags": [
          "collection"
        ]
      }
    },
...
  }
}